suppress a warning

+++ src/linklist.c
@@ -375,7 +375,7 @@ static inline list_entry_t *pick_entry(linked_list_t *list, uint32_t pos)
 
     uint32_t half_length = list->length >> 1;
     MUTEX_LOCK(&list->lock);
-    if (list->cur && abs(list->pos - pos) < half_length) {
+    if (list->cur && (uint32_t)abs(list->pos - pos) < half_length) {
         entry = list->cur;
         if (list->pos != pos) {
             if (list->pos < pos) {
